+++ src/main/kotlin/com/github/itechbear/clion/cpplint/nova/settings/Settings.kt
@@ -15,4 +15,14 @@ object Settings {
     fun get(key: String): String? {
         return instance.getValue(key)
     }
+
+    fun setBoolean(key: String, value: Boolean) {
+        // Store as string because setValue(key, false) UNSETS the key instead of setting it to false!
+        instance.setValue(key, value.toString())
+    }
+
+    fun getBoolean(key: String, defaultValue: Boolean = false): Boolean {
+        val stringValue = instance.getValue(key)
+        return stringValue?.toBoolean() ?: defaultValue
+    }
 }